requirements.txt
flask
python-dotenv
psycopg2-binary
pymongo
bcrypt
PyJWT
requests
flask_jwt_extended
datetime
gunicorn==20.1.0
Flask-CORS